Partial SurfaceEllipse rendering.
Changed the SurfaceEllipse class to maintain an additional parameter
called "theta". This parameter is responsible for the angle between the
first and last interval of the ellipse and allows for partial ellipses
to be drawn. The default value of theta is 360-degrees, meaning that a
full ellipse will be drawn when constructing with the default
parameters.