Cached tile tracking during bulk-download.
Extended on the BulkRetrievalEvent mechanism to allow for callbacks to
occur when tiles are already cached for a specific sector. This logic
was added to the BasicTiledImageLayerBulkDownloader where local or
absent tiles are detected. The BulkRetrievalEvent that is raised for
this is the RETRIEVAL_CACHED event that should be caught when
implementing the BulkRetrievalListener interface.