MeasureTool setPositions() fix.
Made some small changes to the MeasureTool. In particular if the
shape-type was set to SHAPE_LINE or SHAPE_PATH and we seed the
MeasureTool with positions then the shape-type should stay set to what
it was. The old logic caused the MeasureTool to auto-detect the number
of positions and then set the shape-type appropriately, also potentially
arming the measure-tool in the process.