BusyImage moved to separate class.
Moved the BusyImage class that was embeded in the DialogAnnotation class
out to its own separate class (it was also protected previously) as it
is a useful enough class that can be used by clients of this library.
Renamed the class to be BusyImageAnnotation to adhere to correct naming
conventions.