BusyImageAnnotation auto-update control.
Added an auto-update flag to BusyImageAnnotation to allow for enabling
or disabling of auto-updates to the annotation. This is particularly
useful to improve performance and avoid updating during each rendering
cycle. Instead, the doUpdateState() can now be called externally to
flag that an update should occur during the next rendering pass.