SurfaceImage isVisible property.
Modified the SurfaceImage class by adding a new isVisible field that is
used to determine whether the SurfaceImage is visible or not. Added a
getter and setter method for this field. Also, modified the preRender(),
render() and pick() methods to read this field to determine whether the
method bodies should execute or not.