Shapefile borders and coastlines layers.
Added two new XML files called CoastlinesLayer and BordersLayer. These
layers separate the coastlines and country borders into two separate
layers as opposed to the WorldBordersShapefile layer that contains
everyting combined into a single layer. To achieve this, the
corresponding Coastlines and Borders shapefiles were added to the
testData folder. These layers are useful in scenarios where you want to
toggle the country boundaries and coastlines separately, or if you only
want to configure your layers so that either the coastlines or country
boundaries are visible.