Constant field refactoring for various layers.
Moved the default values of LatLonGraticuleLayer, ScaleBarLayer and
SkyColorLayer to public static final variables so that these values can
be used by external clients. This is useful in cases where an external
client might want to toggle between the default value and an internal
value for some of those properties.