DialogAnnotation closing cursor bugfix.
Fixed a bug in the DialogAnnotationController where the cursor would not
reset to the default cursor after closing the DialogAnnotation. The fix
for this bug is to add a call to the setCursor() method in the
onActionPerformed() method to set the cursor to the default-cursor.
Also, the order of the updateCursor() and forwardButtonAnnotation()
method calls was swapped around in the onSelected() method. This is to
ensure that the onActionPerformed() method is called after the
updateCursor() method ensuring that the onActionPerformed() method call
is last so that the default-cursor change persists.