MeasureTool square-placement bugfix.
Fixed a bug in the MeasureTool class where placing a square
measure-shape on the globe resulted in the square being offset by half
the diagonal length in the direction of the center-point. Added
additional logic to check if the square shape is being placed for the
first-time (not being moved) to prevent the unnecessary repositioning of
the square.