SurfaceEllipse location calculation fixes.
Fixed the computeLocations() method in the SurfaceEllipse class.
Firstly, modified the final angle calculation so that theta.radians is
used directly instead of calculating it as i * da which could lead to
rounding errors. Also, changed the actual location calculation to make
use of LatLon.rhumbEndPosition() instead of
LatLon.greatCircleEndPosition(). The reason for this is that the
rhumb-end-position method will ensure that all 3 points (left start
position, center and right end position) are in-line when setting theta
to 180-degrees and using a large radius.