Path dot-drawing offset and limit.
Added two variables to the Path class to control the start and end-index
that is considered when drawing dots for the Path. This allows us to
skip a number of positions before drawing a dot (by setting the offset)
as well as stopping before reaching the last position (by setting the
limit).