MeasureTool ellipse and circle render bugfix.
Fixed a bug in the MeasureTool class where a jagged artefact would
appear if zoomed out far enough and a circle or ellipse is placed on the
globe. The fix for this is to check if the shapeOrientation is small
enough and then to snap it back to zero if this is the case.