FlyToOrbitViewAnimator null-Globe bugfix.
The Globe that is provided to the OnSurfacePositionAnimator's
constructor could, in some situations, be null. This is especially true
during startup when the View.goTo() method is called and the
View.getGlobe() method returns null. As noted in the Javadocs,
View.getGlobe() returns null if called before the first invocation of
View.apply(). So then, instead of getting the Globe, and passing it on,
the OrbitView itself is now passed along as constructor argument where
the OnSurfacePositionAnimator can get the Globe, check if it is null or
not and based on this determine whether it wants to try and get the
endElevation at that point in time. If this fix is not done, the
OnSurfacePositionAnimator could get stuck with a null Globe and will
cause the entire application to break with NullPointerExceptions being
printed out continually.