ShapeEditor utility-line changes for SurfaceText.
Modified the ShapeEditor class by adding an additional line called the
utility-line. This line is used to make the scale control-point for
SurfaceText objects a bit easier to see. The scale control-point was
also moved to now be at 45-degrees east. This is to ensure that the
rotation-line (and control-point) is consistent (pointing at north) with
the rest of the shapes.