AbstractSurfaceShape attribute drawing checks.
Modified the AbstractSurfaceShape class by adding additional
short-circuit logic to skip drawing entirely when both the interior and
exterior should not be drawn. This is done by checking the
isDrawInterior() and isDrawOutline() flags as well as the
getInteriorOpacity() and getOutlineOpacity() values in the
drawGeographic() method before calling the beginDrawing(),
doDrawGeographic() and endDrawing() methods.