ContourLine color and opacity bugfixes.
Refactored the ContourLine class so that it now contains an internal
ShapeAttributes field called lineAttributes that is used in the
makePathsConnected() method instead of creating a new
BasicShapeAttributes instance for each Path that we construct. This also
allows us to simplify the setColor() and setLineWidth() methods
significantly as we don't have to iterate over all Renderables to set
the new color or line-width. Instead we can now set the appropriate
values on the lineAttributes field and the values will propagate through
to the individual lines. By doing this we inadvertently also fixed a bug
where setting the line-width or opacity of the color would not apply to
the individual contour-lines.