Shapefile reduce strictness of content-type check.
Modified the Shapefile and DBaseFile classes by reducing the strictness
of the content-type checks in the validateURLConnection() methods. If we
don't do this and we try to load a shapefile from the classpath via a
URL-string, then the content-type will return as "content/unknown" which
will result in the validation to fail. However, we now handle a
content-type of "content/unknown" in the same manner as an empty
content-type in that we skip the validation check.