Sentinel2 Layer.
Added a new XML configuration file called Sentinel2.xml to the
config/Earth directory. This file provides the configuration for using
the Sentinel 2 terrain imagery data as a layer. The data is pulled from
the tiles.maps.eox.at server. Also, added an entry to the
worldwind.layers.xml file that can be uncommented to add the layer to
the list of WorldWind layers.